CommuteCA2: Standardized Methods for Accessibility Analysis in Canadian Research

The main objective of the *CommuteCA* package is to develop standardized methods for accessibility analysis in research, especially for studies using Statistics Canada surveys. Among the available surveys, we focused our efforts on the 2021 Census of Population, which contain valuable variables for transportation research. This package was created with support from the office of the Research Data Center at McMaster University, the Sherman Centre for Digital Scholarship and Mobilizing Justice.

See also

Author

Maintainer: Bruno Dos Santos dossanb@mcmaster.ca (ORCID)

Authors:

Antonio Paez paezha@mcmaster.ca (ORCID)