Census Tract Land Use (by Mode)

This object contains the census tracts (CT) for all census metropolitan areas and census agglomerations (2021), with land use information regarding the number of labour force (estimated) and number of jobs of each CT. CTs are small, relatively stable geographic areas that usually have a population of fewer than 7,500 persons, based on data from the previous Census of Population Program. They are located in census metropolitan areas (CMAs) and in census agglomerations (CAs) that had a core population of 50,000 or more in the previous census. The labour force and number of jobs in each CT was obtained from the Census of Population. We estimated the labour force by calculating the proportion of trips made by each mode of transport in each census tract, then distributing non-commuters (people who either don't work or work from home) among these transport modes. The number of jobs was obtained applying the weighted sum for cases of respondents mentioning working in the DA. Only cases of CT in accordance with the Research Data Centre vetting rules were released. The methodology to obtain this data set is described in the R markdown 'making-data-v2.Rmd'.

Usage

data(land_use_CT_mode)

Format

A data frame containing 23,970 rows and 9 variables; each row represents a unique CT.

CTUID

Uniquely identifies a census tract. Each CT is assigned a seven‑character numeric "name" (including leading zeros, decimal point and trailing zeros).

CTNAME

Census tract name.

CMAUID

Census Metropolitan Area or Census Agglomeration unique identifier.

CMANAME

Census Metropolitan Area or Census Agglomeration name.

PCD

Uniquely identifies the census division.

PRUID

Uniquely identifies a province or territory.

PwMode

Transportation mode. Possible values: 'Bike', 'Car', 'Transit', and 'Walk.'

labour_force_est_rounded

Estimated number of individuals in labour force (employed or not), rounded to the nearest multiple of 5.

jobs_rounded

Number of jobs in this CT rounded to the nearest multiple of 5.

Source

"The 2021 census tracts areas files as created by the Canadian Census available here accessed May 6th 2024. All variable definitions are based on the definitions included in the census year 2021 boundary file reference guide (Boundary Files, Reference Guide, Second edition, 2021 Census. Statistics Canada Catalogue no. 92-160-G.).